move getDefaultDates
